Added a delegate protocol and property
These delegate methods allow another controller to be notified when the
modal has been displayed and closed. It also provides a way to control
wether the modal should be dismissed or not. This could be used for
determining whether the user has performed any required actions and
preventing them from continuing if not.